refactor: extract ProfileSwitcheri, LanguageSetter, KaniMenu component to separate file
